requirements.txt
Flask==2.2.3
Flask_Cors==3.0.10
numpy==1.24.2
mysql-connector-python==8.0.32
pandas==2.0.0
pdfplumber==0.9.0
PyJWT==2.8.0
pypdf==3.15.0
tabula==1.0.5
tabula_py==2.7.0
python-dotenv==1.0.0
hiredis==2.2.3
redis==4.6.0
